// src/types.ts
export interface Lead {
  id: string;
  name: string;
  email?: string;
  phone?: string;
  timestamp: string;
  status?: 'New' | 'In Progress' | 'Contacted' | 'Closed';
}